+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# MAIN
+# VERSION 0.13.0 - 12 February 2024
 
 - GUI - Store active user in DB rather than in session **(run `rake db:migrate`)**